pyproject.toml
[tool.poetry]
name = "pdf-ppt-classifier"
version = "0.1.0"
description = ""
authors = ["Aaron Wan <baronhw808@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
scikit-learn = "^1.5.1"
pymupdf = "^1.24.9"
pillow = "^10.4.0"
numpy = "^2.0.1"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
